resetructured backend folder and refactored controllers

+++ database-backend/controllers/itemController.js
@@ -1,10 +1,10 @@
 const mongoose = require('mongoose')
-const Items = require('../dbItems')
+const Items = require('../models/dbItems')
 
 // Get Items List
 const getItems = async (req, res) => {
     try {
-        const allItems = await Items.find({}).sort({ createdAt: -1})
+        const allItems = await Items.find({}).sort({ category: 1})
         res.status(200).send(allItems)
     } catch (error) {
         res.status(400).send(error.message)
@@ -13,12 +13,12 @@ const getItems = async (req, res) => {
 
 // Create Items List
 const createItems = async (req, res) => {
-    const dbItems = req.body
+    const {name, category, description, recommendedUses, specialRequirements, numberAvailable, productURL, location} = req.body
     try {
-        const newItems = await Items.create(dbItems)
-        res.status(201).send(newItems)
+        const newItems = await Items.create({name, category, description, recommendedUses, specialRequirements, numberAvailable, productURL, location})
+        res.status(200).send(newItems)
     } catch (error) {
-        res.status(500).send(error.message)
+        res.status(400).send(error.message)
     }
 }
 
@@ -31,14 +31,14 @@ const updateItems = async (req, res) => {
             return res.status(404).send(`There is no Item with id of ${id}`)
         }
         const itemID = {_id: id}
-        const update = { completed : true}
+        const update = { ...req.body}
         const updateItem = await Items.findOneAndUpdate(itemID, update)
         if (!updateItem) {
             return res.status(404).send(`There is no Item with id of ${id}`)
         }
         res.status(200).send(updateItem)
     } catch (error) {
-        res.status(500).send(error.message)
+        res.status(400).send(error.message)
     }
 }
 
@@ -51,9 +51,12 @@ const deleteItems = async (req, res) => {
             return res.status(404).send(`There is no Item with id of ${id}`)
         }
         const deleteItem = await Items.findOneAndDelete({_id: id})
+        if (!deleteItem) {
+            return res.status(404).send(`There is no Item with id of ${id}`)
+        }
         res.status(200).send(deleteItem)
     } catch (error) {
-        res.status(500).send(error.message)
+        res.status(400).send(error.message)
     }
 }
